Reformat and compiling with both versions of CDAWG algo
